Major app-wide fixes: auth, security, UX, dark mode
Critical fixes:
- Payments amount: integer→doublePrecision (DB + schema)
- Real login auth: admin@tesgrup.cz/admin123 (was accepting anything)
- Auth middleware: redirects unauthenticated to /login
- Error handling on cashflow dashboard + prijmy (retry button)
- Delete payments button in Dev. příjmy table

Security:
- Logout button added to sidebar
- Root redirect simplified (middleware handles auth)

UX improvements:
- Portfolio: added print/export button
- Project dashboard: financování mini-widget (vlastní kapitál,
  bankovní úvěr, investorský úvěr, náklady financování)
- Overhead: validation — blocks >100% allocation + duplicate projects
- Dev. příjmy: delete button per payment row

Dark mode:
- Loading spinner contrast fix
- Select option backgrounds
- Tooltip/popover visibility
- Yellow alert/info banner colors

+++ src/app/api/projekty/[projectId]/platby/route.ts
@@ -20,7 +20,7 @@ export async function POST(req: NextRequest, { params }: { params: Promise<{ pro
   if (!saleId || !amount || !paymentDate) {
     return NextResponse.json({ error: 'Chybí saleId, amount nebo paymentDate' }, { status: 400 });
   }
-  const payment = await createPayment({ saleId, projectId, amount: parseInt(amount), paymentDate, label, notes });
+  const payment = await createPayment({ saleId, projectId, amount: parseFloat(amount), paymentDate, label, notes });
   return NextResponse.json(payment, { status: 201 });
 }
 
@@ -28,7 +28,7 @@ export async function PUT(req: NextRequest) {
   const body = await req.json();
   const { id, ...data } = body;
   if (!id) return NextResponse.json({ error: 'Chybí id' }, { status: 400 });
-  if (data.amount) data.amount = parseInt(data.amount);
+  if (data.amount) data.amount = parseFloat(data.amount);
   const payment = await updatePayment(id, data);
   return NextResponse.json(payment);
 }
